+++ modules/accessors/snowflake/lib/index.js
@@ -1,5 +1,9 @@
 import { version }  from '../package.json';
+import { featureCollection, feature, envelope } from '@turf/turf';
 
 export default {
+    featureCollection,
+    feature,
+    envelope,
     version
 };