+++ modules/accessors/snowflake/sql/_SHARE_CREATE.sql
@@ -8,4 +8,5 @@ CREATE SHARE IF NOT EXISTS @@SF_SHARE_PUBLIC@@;
 grant usage on database @@SF_DATABASE@@ to share @@SF_SHARE_PUBLIC@@;
 grant usage on schema @@SF_DATABASE@@.@@SF_SCHEMA@@ to share @@SF_SHARE_PUBLIC@@;
 
+grant usage on function @@SF_PREFIX@@accessors.ST_ENVELOPE(ARRAY) to share @@SF_SHARE_PUBLIC@@;
 grant usage on function @@SF_PREFIX@@accessors.VERSION() to share @@SF_SHARE_PUBLIC@@;