+++ modules/accessors/snowflake/CHANGELOG.md
@@ -8,4 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Create accessors module.
+- Add ST_ENVELOPE function.
 - Add VERSION function.